requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
statsmodels>=0.13.0
matplotlib>=3.5.0
seaborn>=0.11.0
yfinance>=0.1.70
scikit-learn>=1.0.0
jupyter>=1.0.0
pyyaml>=6.0
